Multiple-Disease-Prediction-Webapp/code/PIMA/requirements.txt
numpy
pandas
scikit-learn
xgboost
catboost
optuna
imbalanced-learn
joblib
pyyaml
matplotlib
seaborn
reportlab
category_encoders
scipy
shap
